backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
dotenv==0.9.9
duckdb==1.2.1
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jwt==1.3.1
mibian==0.1.3
motor==3.7.0
numpy==2.2.4
pandas==2.2.3
passlib==1.7.4
psycopg2-binary==2.9.7
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pymongo==4.11.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
wheel==0.45.1
boto3
python-multipart
sqlalchemy
aiogram